requirements.txt

python-telegram-bot==20.0
python-dotenv
aiohttp
nest_asyncio
requests
